+++ lib/shared/core/constants/git_commands.dart
@@ -5,4 +5,5 @@ class GitCommands {
     "--format=%(refname:short)|%(HEAD)",
   ];
   static const List<String> switchToBranch = ["switch"];
+  static const List<String> checkoutBranch = ["checkout", "-b"];
 }