+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - [Branches] List branches.
 - [Branches] Show current branch.
 - [Branches] Switch branch.
+- [Branches] Create and checkout new branch.